+++ attributes/ssh.rb
@@ -1,2 +1,2 @@
 # SSH Allowed Networks (IP/NETMASK)
-default['iptables-nativex']['ssh']['addresses'] = Array.new
+default['iptables-simplyadrian']['ssh']['addresses'] = Array.new