+++ recipes/default.rb
@@ -1,12 +1,12 @@
 #
-# Cookbook Name:: iptables-nativex
+# Cookbook Name:: iptables-simplyadrian
 # Recipe:: default
 #
-# Copyright 2014, NativeX
+# Copyright 2014, simplyadrian
 #
 # All rights reserved - Do Not Redistribute
 #
 
-include_recipe "iptables-nativex::privoxy"
-include_recipe "iptables-nativex::ssh"
-include_recipe "iptables-nativex::web"
+include_recipe "iptables-simplyadrian::privoxy"
+include_recipe "iptables-simplyadrian::ssh"
+include_recipe "iptables-simplyadrian::web"